Created separate file for ReviewListEntry component. Added report functionality to individual review tile. Clicking on report hides clicked review in the database, then re-renders reviews list
