+++ src/components/ratingAndReviewsComponents/RatingAndReviews.jsx
@@ -99,7 +99,7 @@ class RatingAndReviews extends React.Component {
   render() {
     return (
       <div>
-        {console.log(this.state)}
+        {/* {console.log(this.state)} */}
         <hr />
         <div>RATINGS & REVIEWS</div>
         <div>
